Supporting legacy and new input systems for looking, removing prefabs
Prefabs are going to be part of an example project, not the core
package. While I wanted the prefabs to be in the package so people who
add the package can easily drag in a basic character, it's hard to deal
with the fractured input systems nicely given the script references will
break if users don't have the input system enabled for a given prefab.