Adding some of the base movement logic
This is a copy of a bunch of code I wrote for another project, very
roughly yanked out to here. It needs a lot of cleanup and more features
but it gives a good starting point. In particular the CapsuleBody gives
us the basis for a reasonable kinematic player with full movement
control which is nice. The problem now is that it's a very low level
component and, in my other project, required a lot of code around it to
make it usable. I need to figure out how best to split out the stuff
that 90% of games need from the 10% of extra stuff that will be more
game specific.