Move a bunch of code from my high level controller into the CapsuleBody
Obviously still very rough but this moves all the code that was in my
example controller script into a mixture of the CapsuleBody and a new
component for getting movement input.

I might later split up the CapsuleBody from this higher level movement
logic, but at least now it's all in this package as a starting point.