Collapsing all input into a single PlayerInputProvider
Since we lack the ability to easily use interfaces, I don't want my
player to have to have a bunch of tiny input components. So I'm just
merging them together into a single input provider that will give all
the input for all player actions.

As a small perk I now have support for moving with the new input system.
Though I still want to improve the editor/validation of setting up which
actions to use. There are probably some other considerations I need to
think through but for now this is working.