Extracting "player" logic out of CapsuleBody
This restores CapsuleBody to being a "primitive" component in that it
provides some management of, and a wrapper around, a RigidBody and a
CapsuleCollider. It no longer contains any notion of player input,
gravity, velocity etc.

Instead we now have PlayerController which is one way to drive the
CapsuleCollider. My goal for PlayerController (which will likely get
renamed) is that it provides a very reasonable default controller but
is ultimately an optional piece of the system. Games with specific
requirements may want to implement their own version. Or perhaps I split
this up into something more configurable. Either way I expect this area
to have a lot of change since it gets into pretty game specific logic
easily.