Making this not a Unity package manager asset
As far as I can tell packages cannot have optional dependencies which
makes it harder for me to support both the legacy input system and the
new input system. So instead I'm going to work on this repo with the
idea that folks will just drop it into their project directly rather
than using the package manager.

Kind of a bummer, but I'm not going to stress over the package thing
if it's getting in the way.