Making player prefab, removing some serialized component references
While having these all be exposed is very flexible, it also makes the
UI very noisy and can cause issues when moving components and game
objects around. So I'm trying to strip down some of the excess so the
inspector is a little simpler and so there are fewer ways to break
things.