Better type for referencing input actions, including inspector UI
New PlayerInputActionReference gives a much nicer storage of input
actions by storing their IDs rather than their names. This makes it safe
to rename actions without breaking references.

Additionally this adds a property drawer so those fields can be shown
with a nice popup selector populated with the names of actions in the
PlayerInput attached to the same object.