Faking a cylinder cast for ground detection
Raycasts weren't super great because the player fell as soon as their
center went over the edge. Due to how we have the capsule up for step
height they would fall a little bit and then "bump" the edge on the way
down.

I really dislike the way capsule/sphere ground detection works because
it makes the character "round" the corner before falling which feels
very odd. The character should either be on the ground or not; no weird
partial sliding down the edge.

So I figured out I could use a sphere cast but manipulate the math a
little bit to get it to behave like a cylinder. So now the player can
walk out into air a little bit before falling but it meets my binary
criteria for being grounded.

I'll spend some more time on this later to figure out how to allow
players to not be able to go so far over the edge before falling but
this feels like an improvement over the single point raycast.