Shrinking player radius so cylinder ground cast doesn't feel weird
The player was able to go pretty far off a ledge before falling. Rather
than do some complicated hack for collision geometry, I'm just reducing
the radius of the player by a good amount. So far in 16x9 the camera
still doesn't clip into geometry on the corners and it feels much better
in terms of falling off ledges.

I realized the other day (and confirmed in testing) that the cylinder
ground detection does cause stuttering going up ramps so now I have to
figure out how to deal with that...