Turning the camera eye height into, basically, a spring
While some of this is to deal with crouching, this spring system for the
camera height smooths out transitions between the ground and ramp as
well as stairs by compensating for changes in the ground by adjusting
the eye height (relative to the body) and animating it back into place.
This lets us keep our kinematic body spot on while giving the player a
smoother view into the world.