Initial crouch capability
This is pretty rough but I want to continue getting functionality into
place before I do the cleaning up, both of the code and the in-game
behavior. Currently if you let go of crouch in the tunnel things get a
little wonky, but that's fine for now. The code is also a bit of a mess.
Also fine for now.

Eventually the different capabilities will be extracted into simpler
classes and the PlayerController should become more generic. At least
that's the hope so it doesn't stay this big ball of mud forever.