Fix how skin thickness is applied during sweep tests
Rather than treating skin thickness as a forward "bumper" during
movement, we now treat it as a "rewind" for all motion. This still
fulfills the purpose of the skin thickness value (to aid in detecting
collision with either very thin things or surfaces the player is
mathematically planar with) but does so in a way that doesn't distort
the player's collision geometry.

This in turn fixes two issues:

1. The player can no longer clip through the tunnel when coming out of
   crouch.
2. The player does not have a "pop" in position when walking into an
   interior corner and rotating from one face to the other.

Obviously the first one is a bigger deal but the second one was bugging
me. If you slowly turned from one face to another (while walking) you'd
see the position pop at some point as a result of us effectively
distorting the capsule with our sweep test. Now that we apply it as a
rewind instead, we have consistent collision and no more popping. Yay!