Continuing CapsuleBody cleanup, creating body/collider at runtime
This is something I've wanted to do for a bit. Since the CapsuleBody
drives so much of the behavior there's no reason for the user to ever
have to think about the built-in physics components. As such we now
create those components at runtime instead of requiring them to exist in
the editor. This not only simplifies the inspector and slightly reduces
redundancy in the scene/prefab files, but ensures there isn't a
confusing contradiction in configuring the components.

This also forced me to draw the gizmos for the body (since I was relying
on the CapsuleCollider before) so I also took the opportunity to draw a
gizmo to help visualize the cylinder shape used for ground checking.