More cleanup so that the CapsuleCollider is set-and-forget
I wanted to remove all the places where we query the CapsuleCollider for
information since we control all those values. This just required a
couple of extra fields that get set in a few places but now we only ever
put values into CapsuleCollider when resizing and never pull values out.

I did duplicate the setting of those fields but I think in this case I'm
not super worried since the logic for those fields is highly unlikely to
ever change since it describes the fundamental shape. But maybe in the
future I'll refactor out a method to do that.