Replacing all MovePosition with direct modification of position
The change I made before was to have all modifications of the body
position go through the Translate helper which uses
Rigidbody.MovePosition. I did this mostly to make all changes
consistent. However in testing I was able to walk into the obstacle by
walking into the interior corner by the stairs. Upon review I remembered
that MovePosition delays the change for a fixed step so it can do visual
interpolation for any frames rendered in that time. That means if you
call MovePosition and later query the body's position (in the same step)
you won't observe any change! This broke the sweep tests and such.

I think the better fix is to invert the relationship between the body
and controller so that the body calls into the controller in FixedUpdate
and then can accumulate all changes to the position in one go before
applying them, using MovePosition, so we can get interpolation back. But
since that's a bigger architectural change I'm doing this smaller step
first.