Standing up does a real collision check instead of raycast
Raycasts are imperfect because it allows the player to stand up into the
edge of the tunnel and be forced out. Instead we'll use this nicely
renamed helper method to see if the player could stand up and not be
colliding. This ensures the player only stands up if they can do so
without extra movement required to resolve overlaps.

This helper, for more context, is useful for testing hypotheticals with
the player. In this case we're testing just the height but we can also
test positions. One place this is handy (in the future) is something
like mantling where we might want to see if a player would fit on top of
a ledge without colliding at the current or a reduced size. For example,
if you wanted to allow the player to mantle into a vent you would pass
in the position at the top and the crouching height to see if the player
would be able to fit up there.