Adding the ability to lean the camera to see around corners
Most of my character controllers are heavily influenced by the idea of
first person stealth/spy games, such as Neon Struct. One absolute must-
have feature in these games is leaning so the player can see around
corners a little bit without exposing themselves. So naturally I
implemented that.

In this implementation we use an axis for the lean so while generally
this will be binary (keyboards or gamepad shoulder buttons maybe) you
could map it to an analog input for fully analog leaning controls.

Additionally we do sphere casts to ensure the camera can fit where we're
leaning and we nicely clamp both the rotation and distance the camera
moves when there is a collision. This feels nice and lets the player
still lean a little if they can without clipping into walls.

Also, and I can't stress this enough, I'm still just slamming in
features to PlayerController with little regard for overall organization
or optimization at this point. There's one more feature (sliding, as in
Neon Struct or Apex Legends) that I want to put in before I start to
clean up this mess into something a little easier to edit in the
inspector and that will support easily turning features on/off for games
that may want the basic movement but don't want leaning or sliding or
whatever.