Organized some scripts and refactored out player abilities
It's not totally clean and extensible yet but it's getting there. Most
of the data and logic related to the specific abilities now lives
outside of the PlayerController class. This is a starting point toward
making it easier to turn on/off abilities as well as make it easier to
extend with new abilities in the future.