Continuing to work towards a more extensible ability system
Added a new StatefulPlayerAbility for abilities that operate as "states"
like walking, running, etc. Not sure this is exactly the right way to
model the problem but it seems to work OK for now.

In addition the PlayerController now uses the classes themselves as the
state instead of an enum which cleans up a bit of code and lets us use
type parameters for changing the state and leverage virtual methods for
common hooks into state lifetime and relevant events.