Kind of broken right now but a step forward
Working on getting abilities to be entirely generic/polymorphic as far
as PlayerController is concerned so it's easier to make new ones and
add/remove the ones the developer wants from a PlayerController. This
works at runtime but the editor is a super big mess so I want to clean
that up before merging to main.