Continuing to make things more generic
I do think there's an issue with abilities representing states in that
the slide has to force a transition to crouch because if the player
isn't holding crouch then the transition fails. It feels unfortunate to
have two concepts but I do think I need to split "abilities" as
something the player can invoke through input from "states" which are
modal alternations to the state and behavior of the player.