Exposing body and input properties so abilities can handle more logic
Basically I wanted to inline a lot of the work of some properties so
that the controller doesn't need to have properties that speak to the
abilities themselves. Instead the abilities can just speak directly to
the input system to get the values they need.

Obviously the input system still knows about the various abilities but
I'm not really sure of a way around that without having each ability
handle input on its own, which might work but might require a lot of
crufty code if we care to support both input systems.