Changed how blocking works, abilities can update when not active
Instead of just skipping all abilities under the active blocking one, we
now deactivate all abilities below the active one. I'm sure there's some
fun bugs in here, but it helps us fix leaning.

So leaning was broken in an odd way where you could lean to the side,
then start running, and while running you'd be stuck in your lean. This
is because leaning is higher priority (and blocking) so lean stopped
updating. The maybe-not-quite-super-elegant solution I've added is the
ability for abilities to opt in to being updated when not active. I'm
not super crazy about this but it allows the lean to always update but
then requires that ability to manually check if the ability is active or
not.

I also made a bunch of methods into properties because I like seeing
fewer ()'s in the code.