Fixing canStandUp to not return false for walls
When I changed crouch to activate anytime the player can't stand up the
idea was to auto crouch after sliding if you slide into a tunnel. It was
fixing a bug where the player automatically stood up inside the tunnel.
However I found out soon after that running into some walls would cause
the player to auto-crouch. The issue is in how we checked for whether
the player can stand up using a full capsule check.

I think a sphere cast might be the better approach, but this at least
fixes the issue for now.