Dramatically simplifying abilities editor, fixing a big bug
Turns out adding didn't work because I needed to call
ApplyModifiedProperties. My bad.

Also I was able to really reduce the code in the editor by just relying
on the built in rendering of property fields to do a lot of the work for
me. The items in the list now have a more "boring" label on them but it
seems to work for me and I like being able to remove a ton of code so
that this thing isn't so complicated and error prone.