Trying out SerializedObject because managed references are kind of bad
Managed references are stored with assembly name and class name info
which makes it hard to rename anything and not break stuff. On the other
hand, normal references in Unity use a GUID system that doesn't depend
on the class or assembly name. However to use SerializedObject we have
to make files which is kind of ugly. But I wanted to try it out and see
if I like it.