Implementing new property drawer for player abilities
Even with SerializedObjects for abilities, we need to be able to easily
reorder them because the abilities list is prioritized. This is the most
basic reorderable list implementation, but I might add some additional
features later to allow editing the abilities inline, one of the main
benefits of the managed references previously used.