Extracting an interface for the player controller
This is a good bit of cleanup I've been wanting to do so that the
abilities aren't given full access to the entire component/gameobject
for the player but instead are restricted to an interface. This not only
makes it easier to reason about the things abilities can do, but also
ensures that we can use the same abilities with any other controller
implementation, provided it can implement the same interface.

The interface will likely shift over time but this is a nice separation.