CapsuleBody now has option for whether to treat feet as a cylinder
For games that are concerned about precise ledges such as stealth games
or games in more urban settings, cylinders are pretty great for feet.
The player has an absolute point where they fall but until then they are
grounded on the surface. However for games where that isn't a concern
such as games with more organic terrain, the cylinder can create some
odd collisions and a capsule might be more appropriate.

So now you can simply determine whether or not you want a cylinder for
ground checks. And the visualization gizmos update appropriately.