Trying out this idea of "intents" to drive inputs
The general idea here is to consider things in terms of player intent
rather than raw input. While the current intents are super basic pass-
throughs, imagine a more interesting crouch intent. It might support two
different input actions where one must be held to crouch while the other
acts a toggle. A single intent can wrap that all up and expose a single
boolean of "does the player want to be crouched".

That said I'm not sure if this is really that clean/simple or just more
verbose, but I'm willing to give it a run.