Changing how min/max pitch for looking is specified to be nicer to edit
Min/max pitch values aren't very intuitive since it isn't clear which
one is "up" (it was the min pitch which feels odd at least to me). It's
also weird to specify a negative value for looking up. So now both max
angles are positive values and more clearly named.