Scrapping the intent system and moving look input onto shared object
I think the "intent" idea is an interesting design but it also feels a
bit heavyweight requiring a bunch of small classes that don't add up to
anything super amazing. While the monolith input interface doesn't feel
right either when considering game-specific abilities (and wanting to
keep the interface in the package) it feels better for now.