Fixing ground check for capsule body to use step height correctly
The sphere cast was starting from the capsule's center which is
definitely not correct and was allowing my character with a step height
of 0.35 to be able to walk up a 1 meter step. Now we start the sphere
cast from the correct place which, as you might guess, makes things work
as one would expect.