Bringing back the "intent" concept but as component interfaces
I genuinely think that having interfaces express player intents is a
cleaner abstraction for the abilities than directly exposing input. This
not only provides an abstraction behind which you can implement
different control schemes, for example crouch as a toggle versus crouch
as a button hold, but also ensures we don't do things like having the
slide ability checking the crouch input. Instead the ISlideIntent
interface provides a "wantsToSlide" value that can be mapped to crouch
but doesn't have to be.

Additionally this once again makes it possible to support new abilities
with new inputs without requiring a single monolithic interface to be
updated with the new inputs.