Adding crouch toggle support via second action
This is an approach I personally really like in games. Rather than
having some runtime user option for whether or not the crouch button is
a crouch, give the player two input bindings so they can use whichever
makes sense in the moment. If I just need to crouch for a moment, I can
hold a button. If I plan to stay crouched I can toggle.