Removing the layer mask from the sphere cast which isn't necessary
Since our sphere starts inside the player capsule we don't need to worry
about masking the collision. Also this approach wasn't 100% reliable
anyway as it assumes that the player is the only thing on the layer
which might not be the case. If we need to add this back a more proper
solution is to move the player onto the IgnoreRaycasts layer and mask
off that in our cast.