Moving code out of package and back into Assets
I don't really think I like the whole package system. I think for most
people it's easier to just drop the folder into Assets and treat it like
the rest of their code. It's not a big distinction really but this also
will play nicer with the Asset Store if I decide to put the package up
there.