Removing all the ScriptableObject ability configuration stuff
The scriptable objects seemed like they might be useful but in practice
it feels like a lot of extra boilerplate code and asset files. With this
new approach the controller just exposes the set of abilities as
serialized fields making editing of the values much simpler.

One big ramification here is that the controller's ability set is no
longer configurable in the editor. A code change is required in order to
add, remove, or re-order the abilities. I think this is fine and am OK
with requiring some code changes in order to keep things more simpler.