First pass at an ability to limit the slope a player can walk on
This is to try and address issue #1 in a way that is game configurable.
There's no single best way to handle steep slopes in games because each
game has different restrictions. A game that doesn't allow the player
to reach the top of the slope might simply stop the player from going up
it at all, whereas games that allow the player to be up top likely want
something more like this where the player slides down automatically.

I might spend more time on this, but I might wait until I'm working on a
real game that has this need so I can be driving it from experience,
rather than just what I think would make sense. But I'm glad to have
this here as an example approach.